+++ src/main/java/domain/amount/InvestmentAmount.java
@@ -1,9 +1,11 @@
 package domain.amount;
 
+import java.math.BigDecimal;
+
 import domain.interest_rate.InterestRate;
 
 public interface InvestmentAmount {
 	double calAnnualInterest(InterestRate interestRate);
 
-	double calMonthlyInterest(InterestRate interestRate);
+	BigDecimal calMonthlyInterest(InterestRate interestRate);
 }
